Check active record direct descendants only.

+++ features/support/env.rb
@@ -6,17 +6,18 @@ Bundler.setup
 
 if File.directory?("test/rails_app")
   Dir.chdir("test/rails_app") do
-    raise unless system("rm -f db/migrate/*create_invalid_records.rb && rake db:migrate:reset")
+    raise unless system("rm -f db/migrate/*create_invalid_records.rb && rake db:drop db:create db:migrate")
   end
 end
 
 After do
   # Reset DB!
-  tables = ['users', 'categories', 'posts']
+  tables = ['categories', 'emails', 'invalid_records', 'posts', 'users']
   conn = ActiveRecord::Base.connection
-  tables << 'invalid_records' if conn.table_exists?('invalid_records')
   tables.each do |table|
-    conn.execute("DELETE FROM '#{table}'")
-    conn.execute("DELETE FROM sqlite_sequence WHERE name='#{table}'")
+    if conn.table_exists?(table)
+      conn.execute("DELETE FROM '#{table}'")
+      conn.execute("DELETE FROM sqlite_sequence WHERE name='#{table}'")
+    end
   end
 end
